Changed aave pool address contract(v2) & found bug in calculate value in cds

+++ Blockchian/contracts/interface/ITreasury.sol
@@ -19,7 +19,7 @@ interface ITreasury{
         uint64 ethPriceAtWithdraw;
         uint64 withdrawTime;
         uint128 aBondTokensAmount;
-        uint64 strikePrice;
+        uint128 strikePrice;
         uint128 optionFees;
         uint256 burnedAmint;
         uint64 externalProtocolCount;
